feat(build-release): Prepare & publish license & notice artifact (#6)
Our Android Service release needs to include license terms and third party notices. This PR updates the release build to include the terms and notices in a build artifact. It also adds a name to the apk release artifact for clarity. These will both be consumed in a release pipeline.