feat(android-setup): add accessibility-insights-for-android-service-bin NPM wrapper (#18)
#### Description of changes

This PR adds an `/npm-wrapper` folder that implements an NPM package named `accessibility-insights-for-android-service-bin` to act as a thin wrapper around the APK for the repo, and updates `release-build.yaml` to combine the signed APK with the `npm-wrapper` content to produce a packed `.tar.gz` version of the package for publishing in a release pipeline.

See `/npm-wrapper/README.md` for an overview of the proposed API for the package (it just bundles the APK+license+notice and exports constants containing the paths of the APK+notice)

Since I had to update the release build to repackage assets into the NPM package anyway, I also updated the release build to perform the zip packaging step for the github release artifact; this should enable us to reduce the release build stage for github publishing from 5 tasks to 1, reducing the amount of build process that happens outside of source control.

Collaborators can see a [trial release build](https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=11193&view=results) against this branch; note especially the `npm-wrapper` and `github-release-archive` artifacts it produces. The `npm-wrapper` artifact contains an `accessibility-insights-for-android-service-bin.tar.gz` file which you can test locally by downloading and running `yarn add file:C:\path\to\downloads\accessibility-insights-for-android-service-bin.tgz` from a test node app; I verified that the typescript code from the wrapper's README file successfully builds/typechecks under default typescript settings and that it produces the expected runtime output.

After this PR is merged, I'll be making corresponding updates to the release build to simplify the existing stage using the new `github-release-archive` artifact. Then, I'll add a new variable group hooked up to our production keyvault to retrieve the NPM publishing credential and add a second stage to the release pipeline that uses that variable group and the `npm-wrapper` artifact to publish the wrapper package.

Technically, this will break the existing release briefly until I make the update; I think this will be quick enough and low enough risk to not be worth doing a staged rollout (we can just revert this PR if for some reason it doesn't work out and we need to do a quick release in the old method).

#### Pull request checklist

- [x] Addresses an existing issue: 1724321
- [n/a] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from `AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).