chore: update gradle and migrate to single-lockfile (#34)
#### Description of changes

This PR performs the following coupled updates:

* Gradle 5.4.1 -> 6.6 (latest stable)
* Android build tools Gradle plugin 3.5.3 -> 4.0.1 (latest stable)
* Android SDK build tools 28.0.3 -> (allow gradle build tools plugin to choose, per the new version's recommendation)

The main benefit up updating to Gradle 6.6 is that it supports using one lockfile per project instead of one lockfile per gradle build task. This is much easier to review, more consistent with our other projects, and removes the possibility of different targets using out of sync versions. We're currently enabling a feature flag in `settings.gradle` to enable this; it is slated to become the default in Gradle 7.

Dependabot doesn't understand gradle lockfiles regardless of whether the flag is on and Component Governance does understand them regardless of whether the flag is on.

#### Pull request checklist

<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Addresses an existing issue: #0000
- [n/a] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from `AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).