fix(web-compliance): include missing dependencies in (manual) NOTICE.html (#45)
#### Description of changes

In an earlier PR, we added a task to have Component Governance automatically generate NOTICE.html files for us. Unfortunately, there are several problems that block this approach from being accurate:

1. We have some cases where we use code that Android Studio generated, which means we need an attribution for The Android Open Source Project which Gradle/Component Governance doesn't know about
2. Component Governance will only include "supported components" in NOTICE files, which excludes anything from a cgmanifest file and notably excludes `androidx.annotation`, which is a Maven dependency not present in Maven Central
3. Component Governance doesn't have a deterministic way to tell whether a given dependency is a devDependency or not, so it is conservative and always assumes "real dependency". Unfortunately, this means that every time a devDependency is bumped by dependabot, we need to remember to go manually configure CG to omit the new version, or it will start showing up in NOTICE files

We can sort of work around 1 with a cgmanifest file, but it immediately runs into 2. We don't have workarounds for 2 and 3. Until we do, we have to go back to manually curating NOTICE.html.

This PR does that, but extends and automates the use of the `list-release-dependencies` script into something that verifies that the lockfile and NOTICE.html are in sync as a condition of PR builds passing, to avoid the case where Dependabot updates a version of something important and we forget to update NOTICE.html.

It also adds the necessary copyright header for AOSP to the file in question that triggers us to include it.

#### Pull request checklist

<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Addresses an existing issue: #0000
- [n/a] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from `AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).