chore: Allow non-blocking requests (precursor to long polling API) (#59)
Add an isBlockingRequest method to the RequestFulfiller interface, so each type of request can be either blocking or non-blocking. The ServerThread will block on the ResponseThread only for RequestFulfiller implementations that specify that they're blocking. When we add the RequestFulfiller implementation for focus changes, we'll add it as a non-blocking implementation.