feat(google-rules-mvp-service): Add Google ATFA dependency + supporting changes (#101)
#### Details
This PR adds a dependency for Google's [Accessibility Test Framework for Android](https://github.com/google/Accessibility-Test-Framework-for-Android) (ATFA). To support this, it updates build.gradle to look for that dependency in Google's repository. It also adds a dependency on Guava, which is necessary to access the results of ATFA. NOTICE.html was updated with the new dependencies and verify-notice-contents.js was updated to accept license names with whitespace in them.

##### Motivation
Paving the way for upcoming work to run and return ATFA checks.

##### Context
This could have been a part of whatever PR first uses ATFA, but we wanted to break things up into smaller pieces. 

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Addresses an existing issue: #0000
- [n/a] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from `AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).