feat(google-rules-mvp-service): Introduce ResultSerializer and use in ResultRequestFulfiller (#102)
#### Details
In preparation for adding ATFA results, this PR introduces `ResultSerializer` and uses it to generate the response string in `ResultRequestFulfiller`. This has no functional impact--for now, `ResultSerializer` is just a wrapper around `AxeResult` that calls `toJson()`.

##### Motivation
Ultimately, we need a way to return AxeResults, ATFA results, and ATFA rules serialized alongside each other. This PR introduces the layer of abstraction where this serialization will take place.

##### Context
Out of scope: adding any new results to the response of `ResultRequestFulfiller`. Future PRs will expand upon `ResultSerializer` such that it returns additional data.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Addresses an existing issue: #0000
- [x] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from `AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).