feat(google-rules-mvp-service): Create ATFAScanner (#104)
Details
This PR creates the ATFAScanner, which uses the Accessibility Test Framework for Android, and the ATFAScannerFactory. Like the existing AxeScanner, the ATFAScanner runs the rules (ATFA instead of axe, in this case) on a given root node and then returns the results.

Motivation
This is a part of the feature to add ATFA to AI for Android Service.

Context
Created an instance of ATFAScanner in AccessibilityInsightsForAndroidService.java, but adding it to the ResponseThreadFactory and RequestHandlerFactory will be done in a future PR, most likely when updating the API.

It also filters the results to return only those with type ERROR, INFO, WARNING, RESOLVED, and NOT_RUN. This can be changed in the future, if we decide that we want to return other types of results. Currently, the only other available result type is SUPPRESSED, but those types of results are not added in the rules that we are using.