feat(google-rules-mvp-service): Run, serialize, and return ATFA results (#110)
#### Details
Previous PRs introduced `ATFAScanner`, `ATFAResultSerializer`, and `ResultSerializer` (#104, #108, and #102). This PR puts them together to run, serialize, and return ATFA results. To accomplish this, `ResultsContainer` is introduced and `ResultSerializer` is refactored to serialize it. 

##### Motivation
Now ATFA rules are run and results are returned!

##### Context
A future PR will add ATFARules to the return object, as well. 

Note that these changes will require updates on the AI-Android client receiving end.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Addresses an existing issue: #0000
- [x] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from `AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).