chore: make compiler/lint warnings fail the build (#121)
#### Details

Before this PR, build warnings and lint warnings were being emitted on stdout during builds, but weren't actually blocking the build from succeeding, so they were being ignored and just producing noise.

This PR enables build, test, and compiler warnings to show up as build failures and fixes most pre-existing warnings. This reduces several bits of noise during builds (eg, the "try rebuilding with -Xlint:deprecated" messages and the 6 lines of red WARNING text you get in Android Studio when running tests there).

I've also set up the build to upload a `lint-results` artifact containing lint results (which were previously produced but discarded)

##### Motivation

Avoid noise, fix issues preemptively

##### Context

I included the necessary `.idea` files to run from Android Studio 4.2.2 (latest stable) as part of the PR since I couldn't validate locally otherwise.

The only class of warning which I think would be nice for us to enable but which I didn't include as part of this PR is the `UnknownNullness` lint check, which is similar to strictNullChecks in web; it is a lint which essentially flags any method parameter/etc which has neither a `@NotNull` nor a `@Nullable` annotation. I think being explicit about this is a good idea and approve of the lint rule, but there are ~250 instances of the warning in the codebase so I wanted to leave adding them for a separate PR.

I also intentionally left fixes for the quieted `OldTargetApi` and `GradleDependency` warnings out-of-scope (we'll want to separately investigate why dependabot isn't picking up the old gradle deps)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Addresses an existing issue: #0000
- [x] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from `AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).