chore: Update lockfile after running gradlew fastpass (#196)
#### Details

On a separate PR, I ran `gradlew fastpass`, and it updated the lockfile.
I tried running it on just main, and it also updated the lockfile. I'm
not sure if this is a recent side effect of running on the latest
version of Android Studio or if this is just a step that got missed in a
previous dependabot PR.

##### Motivation

Bring the code into conformance with our PR steps

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [n/a] Addresses an existing issue:
- [n/a] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from
`AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).