chore: Workaround for mockito #2860 (#199)
#### Details

Due to https://github.com/mockito/mockito/issues/2860, Dependbot has
been unable to update to any version of `mockito-inline` since version
4.6.1. The failed attempts are as follows:

Version | Failed PR
--- | ---
4.7.0 | #169 
4.8.0 | #174
4.8.1 | #176
4.9.0 | #182
4.11.0 | #198 

In response to https://github.com/mockito/mockito/issues/2860, a
suggestion was made to use an `ArgumentCaptor` to capture the arguments
and the `RETURNS_SELF` parameter for the builder. This PR makes that
change, so we're now capturing arguments using the construct that is
intended for that purpose. The mockito team has not committed to fixing
this issue and may even be leaning toward not fixing it, since a more
officially supported option exists to do what we're doing.

I've applied these changes locally to #198 and confirmed that the tests
complete as expected.

##### Motivation

Allow newer versions of mockito

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [n/a] Addresses an existing issue:
- [n/a] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from
`AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).