chore: Remove line that Android Studio keeps on removing (#197)
#### Details

When editing the project in the latest Windows version of Android
Studio, it silently removes the `resolveModulePerSourceSet` line on
save. If Android Studio thinks that it's not needed, then I'm inclined
to trust the tool's judgment and remove it once and for all. The
alternative is to ask every time a change is made via the IDE.

##### Motivation

Minimize confusion when using the IDE.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [n/a] Addresses an existing issue
- [n/a] Added/updated relevant unit test(s)
- [x] Ran `./gradlew fastpass` from
`AccessibilityInsightsForAndroidService`
- [x] PR title _AND_ final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`).